@@ -27,91 +27,80 @@
   },
   "homepage": "https://github.com/OpusCapita/onboarding/blob/master/README.md",
   "dependencies": {
-    "@opuscapita/react-loaders": "^0.1.8",
-    "@opuscapita/react-menus": "^1.0.26",
-    "assets-webpack-plugin": "^3.5.1",
+    "@opuscapita/i18n": "^1.1.0",
+    "@opuscapita/react-loaders": "^0.1.9",
+    "@opuscapita/react-navigation": "^0.0.30",
     "async": "^2.5.0",
     "babel-runtime": "^6.23.0",
-    "bluebird": "^3.5.0",
-    "express": "^4.15.3",
+    "bluebird": "^3.5.1",
+    "deep-equal": "^1.0.1",
+    "express": "^4.16.2",
     "express-handlebars": "^3.0.0",
     "handlebars": "^4.0.10",
     "history": "^3.3.0",
-    "jsdoc-to-markdown": "^3.0.0",
     "lodash": "^4.17.4",
-    "moment": "^2.18.1",
-    "node-schedule": "^1.2.3",
-    "nodemon": "^1.11.0",
-    "ocbesbn-blob-client": "^1.0.16",
-    "ocbesbn-bouncer": "^0.2.11",
+    "node-schedule": "^1.2.5",
+    "ocbesbn-blob-client": "^1.0.19",
     "ocbesbn-config": "^2.2.11",
-    "ocbesbn-db-init": "^1.1.10",
+    "ocbesbn-db-init": "^1.1.11",
     "ocbesbn-logger": "^1.0.5",
     "ocbesbn-redis-events": "^1.0.5",
-    "ocbesbn-service-client": "^1.1.3",
-    "ocbesbn-web-init": "^2.2.8",
-    "opuscapita-i18n": "^1.0.12",
+    "ocbesbn-service-client": "^1.1.4",
+    "ocbesbn-web-init": "^2.2.9",
     "papaparse": "^4.3.3",
-    "react": "15.4.2",
-    "react-async-component": "^1.0.0-beta.3",
-    "react-bootstrap": "^0.31.3",
-    "react-bootstrap-table": "^4.0.0",
+    "prop-types": "^15.6.0",
+    "raw-loader": "^0.5.1",
+    "react": "^15.4.0",
     "react-clipboard.js": "^1.1.2",
-    "react-dom": "15.4.2",
+    "react-dom": "^15.4.2",
     "react-dropzone": "^3.13.3",
-    "react-intl": "^2.4.0",
     "react-notification-system": "^0.2.14",
-    "react-redux": "^5.0.5",
-    "react-router": "^3.0.2",
-    "recharts": "^0.22.4",
+    "react-router": "^3.2.0",
+    "recharts": "^1.0.0-alpha.6",
     "redis": "^2.7.1",
-    "redux": "^3.7.1",
-    "redux-form": "^7.0.3",
-    "redux-thunk": "^2.2.0",
     "rimraf": "^2.6.2",
     "sequelize": "^3.30.4",
-    "sequelize-cli": "^2.7.0",
-    "sequelize-fixtures": "^0.6.0",
     "superagent": "^2.3.0",
     "superagent-bluebird-promise": "^4.2.0",
     "validate.js": "^0.11.1",
-    "webpack-bundle-analyzer": "^2.8.2",
-    "webpack-md5-hash": "0.0.5",
-    "webpack-merge": "4.1.0",
     "xlsx": "^0.10.6"
   },
   "devDependencies": {
-    "babel-cli": "6.24.1",
-    "babel-core": "6.25.0",
-    "babel-eslint": "7.2.3",
-    "babel-jest": "20.0.3",
-    "babel-loader": "7.1.1",
-    "babel-plugin-istanbul": "4.1.4",
-    "babel-plugin-lodash": "3.2.11",
-    "babel-plugin-transform-decorators-legacy": "1.3.4",
-    "babel-plugin-transform-object-assign": "6.22.0",
-    "babel-plugin-transform-runtime": "6.23.0",
-    "babel-plugin-webpack-loaders": "0.9.0",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^7.2.3",
+    "babel-jest": "^20.0.3",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-istanbul": "^4.1.5",
+    "babel-plugin-lodash": "^3.2.11",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-object-assign": "^6.22.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-plugin-webpack-loaders": "^0.9.0",
     "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "1.6.0",
-    "babel-preset-react": "6.24.1",
-    "babel-register": "6.24.1",
-    "chai": "4.0.2",
-    "css-loader": "0.28.4",
-    "eslint": "4.1.1",
-    "eslint-config-opuscapita": "1.0.7",
-    "eslint-plugin-import": "2.7.0",
-    "eslint-plugin-jsx-a11y": "6.0.2",
-    "eslint-plugin-react": "7.1.0",
-    "faker": "4.1.0",
-    "jsdoc-to-markdown": "3.0.0",
-    "json-loader": "0.5.4",
-    "mocha": "3.4.2",
-    "mocha-junit-reporter": "1.13.0",
-    "nyc": "11.0.3",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-register": "^6.26.0",
+    "chai": "^4.1.2",
+    "css-loader": "^0.28.7",
+    "eslint": "^4.8.0",
+    "eslint-config-opuscapita": "^1.0.9",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-react": "^7.4.0",
+    "faker": "^4.1.0",
+    "jsdoc-to-markdown": "^3.0.0",
+    "json-loader": "^0.5.7",
+    "mocha": "^3.5.3",
+    "mocha-junit-reporter": "^1.15.0",
+    "nyc": "^11.2.1",
     "raml-to-markdown": "^1.1.7",
-    "sequelize-to-markdown": "1.1.13",
-    "style-loader": "0.18.2",
-    "webpack": "3.1.0"
+    "sequelize-to-markdown": "^1.1.13",
+    "style-loader": "^0.18.2",
+    "webpack": "^3.7.1",
+    "webpack-md5-hash": "0.0.5",
+    "webpack-merge": "^4.1.0",
+    "nodemon": "^1.12.1",
+    "assets-webpack-plugin": "^3.5.1"
   }
 }
