@@ -1,98 +1,181 @@
-import React, { Component } from 'react';
-import ajax from 'superagent-bluebird-promise';
+import React from 'react';
+import PropTypes from 'prop-types';
+import Api from './api';
 import serviceComponent from '@opuscapita/react-loaders/lib/serviceComponent';
-import { Popover, OverlayTrigger, Button } from 'react-bootstrap';
-import FileManager from './FileManager.react';
 import Dropzone from 'react-dropzone';
-import ModalDialog from '../common/ModalDialog.react';
+import translations from './i18n';
+import { ContextComponent, ModalDialog } from '../common';
+import TemplatePreview from './TemplatePreview.react';
+import ClipboardButton from 'react-clipboard.js';
 import validator from 'validate.js';
+import extend from 'extend'
 
-class TemplateForm extends Component
+const templateFields = {
+    en : require(`./data/templateFields.en.json`),
+    de : require(`./data/templateFields.de.json`)
+};
+
+class TemplateForm extends ContextComponent
 {
     static propTypes = {
-        tenantId : React.PropTypes.string.isRequired,
-        template : React.PropTypes.object,
-        filesDirectory : React.PropTypes.string,
-        onCreate : React.PropTypes.func,
-        onUpdate : React.PropTypes.func,
-        onCancel : React.PropTypes.func,
-        allowCreate : React.PropTypes.bool.isRequired,
-        allowUpdate : React.PropTypes.bool.isRequired,
-        allowCancel : React.PropTypes.bool.isRequired
+        customerId : PropTypes.string.isRequired,
+        templateFileDirectory : PropTypes.string.isRequired,
+        templateId : PropTypes.number.isRequired,
+        type : PropTypes.string.isRequired,
+        allowCopy : PropTypes.bool.isRequired,
     }
 
     static defaultProps = {
-        onCreate : () => { },
-        onUpdate : () => { },
-        onCancel : () => { },
-        allowCreate : true,
-        allowUpdate : true,
-        allowCancel : true
+        templateId : 0,
+        type : '',
+        allowCopy : true
     }
 
     static contextTypes = {
-        showNotification : React.PropTypes.func.isRequired,
-        hideNotification : React.PropTypes.func.isRequired
+        showNotification : PropTypes.func.isRequired,
+        hideNotification : PropTypes.func.isRequired,
+        i18n : PropTypes.object.isRequired,
+        locale : PropTypes.string.isRequired
     }
 
-    templateFields = [{
-        key : 'contact.firstName',
-        description : 'Contact first name.'
-    }, {
-        key : 'contact.lastName',
-        description : 'Contact last name.'
-    }]
-    .sort(item => item.key);
+    static emptyFormItem = {
+        id : 0,
+        name : '',
+        content : '',
+        languageId : '',
+        countryId : '',
+        type : ''
+    }
 
-    constructor(props)
+    constructor(props, context)
     {
         super(props);
 
-        this.state = {
-            id : null,
-            name : '',
-            content : '',
-            language : '',
-            country : '',
-            logoFile : '',
-            headerFile : '',
-            isNew : true,
+        context.i18n.register('TemplateForm', translations);
+
+        const basicState = {
+            templateFileDirectory : this.makePathDirectory(props.templateFileDirectory),
+            filesDirectory : '',
+            templateId : props.templateId || '',
+            propsType : props.type,
+            type : props.type,
+            templates : [ ],
+            files : [ ],
             errors : { }
         }
-    }
 
-    componentWillMount()
-    {
+        this.state = extend(false, { }, TemplateForm.emptyFormItem, basicState);
+        this.formChanged = false;
+        this.preview = null;
+
         const serviceRegistry = (service) => ({ url: '/isodata' });
         this.LanguageField = serviceComponent({ serviceRegistry, serviceName: 'isodata' , moduleName: 'isodata-languages', jsFileName: 'languages-bundle' });
         this.CountryField = serviceComponent({ serviceRegistry, serviceName: 'isodata' , moduleName: 'isodata-countries', jsFileName: 'countries-bundle' });
     }
 
-    loadFiles()
+    componentDidMount()
     {
-        return ajax.get(`/blob/api/${this.props.tenantId}/files/${this.props.filesDirectory}`)
-            .then(response => JSON.parse(response.text));
+        this.reload();
     }
 
-    loadTemplate()
+    componentWillReceiveProps(nextPops, nextContext)
     {
-        return ajax.get(`/blob/api/${this.props.tenantId}/files/${this.props.template.path}`)
-            .then(response => response.text);
+        const propsChanged = Object.keys(nextPops).reduce((all, key) => all || nextPops[key] !== this.props[key], false);
+
+        if(propsChanged)
+            this.setState(extend(true, { propsType : nextPops.type }, this.state, props));
+    }
+
+    reload()
+    {
+        return this.loadTemplate();
+    }
+
+    makePathDirectory(path)
+    {
+        return this.makePathAbsolute(path.endsWith('/') ? path : path + '/');
+    }
+
+    makePathAbsolute(path)
+    {
+        return path.startsWith('/') ? path : '/' + path;
+    }
+
+    getFilesDirectory(templateId)
+    {
+        templateId = templateId || this.state.id;
+        return `${this.state.templateFileDirectory}${templateId}`;
+    }
+
+    loadTemplate(templateId)
+    {
+        templateId = templateId || this.state.id;
+
+        if(templateId)
+        {
+            const filesDirectory = this.getFilesDirectory();
+            this.setState({ filesDirectory, templateId });
+
+            return this.loadTemplates().then(() => this.loadFiles()).then(() =>
+            {
+                return Api.getTemplate(this.props.customerId, templateId)
+                    .then(item => this.putItemToState(item))
+                    .then(() => true)
+                    .catch(e => this.context.showNotification(e.message, 'error', 10));
+
+                return Promise.resolve(true);
+            });
+        }
+        else
+        {
+            this.clearForm();
+            return this.loadTemplates().then(() => false);
+        }
+    }
+
+    loadTemplates()
+    {
+        return Api.getTemplates(this.props.customerId)
+            .then(templates => this.setState({ templates }))
+            .catch(e => this.context.showNotification(e.message, 'error', 10));
+    }
+
+    loadFiles()
+    {
+        if(this.state.id)
+        {
+            const filesDirectory = this.getFilesDirectory(this.state.id);
+
+            return Api.getFiles(this.props.customerId, filesDirectory)
+                .then(files => this.setState({ files }))
+                .catch(e => this.context.showNotification(e.message, 'error', 10));
+        }
+
+        return Promise.resolve(false);
     }
 
     handleOnChange(e, fieldName)
     {
-        this.setState({ [fieldName]: e.target.value });
+        this.formChanged = true;
+
+        if(typeof e === 'object')
+            this.setState({ [fieldName]: e.target.value });
+        else
+            this.setState({ [fieldName]: e });
     }
 
-    showFileSelector(forField)
+    applyTemplate(templateId)
     {
-        this.setState({ showFileSelector : forField });
-    }
+        const copyTemplateContent = () =>
+        {
+            this.handleOnChange(templateId, 'templateId');
+            this.loadTemplate(templateId).then(() => this.setState({ id : null, filesDirectory : null }));
+        }
 
-    hideFileSelector()
-    {
-        this.setState({ showFileSelector : false });
+        if(this.state.name || this.state.content)
+            this.showOverwriteTemplateDialog(copyTemplateContent, () => null);
+        else
+            copyTemplateContent();
     }
 
     handleFileSelection(forField, value)
@@ -109,7 +192,7 @@ class TemplateForm extends Component
 
             reader.onload = (e) =>
             {
-                this.context.showNotification('File uploaded.', 'success');
+                this.context.showNotification(this.context.i18n.getMessage('TemplateForm.uploadTemplate.notification.success'), 'success');
                 this.setState({ content : e.target.result });
             }
 
@@ -146,14 +229,23 @@ class TemplateForm extends Component
             this.state.overwriteOnCancel();
     }
 
-    validateItem(item)
+    validateForm(item)
     {
-        validations = {
+        const i18n = this.context.i18n;
+
+        const validations = {
+            type : {
+                presence : { message : i18n.getMessage('TemplateForm.validation.type.isRequired') },
+            },
             name : {
-                presence : true,
+                presence : { message : i18n.getMessage('TemplateForm.validation.name.isRequired') },
                 format : {
-                    pattern: /[a-z0-9-]/
+                    pattern : /^[a-z0-9-]*$/,
+                    message : i18n.getMessage('TemplateForm.validation.name.format')
                 }
+            },
+            content : {
+                presence : { message : i18n.getMessage('TemplateForm.validation.content.isRequired') }
             }
         }
 
@@ -164,171 +256,258 @@ class TemplateForm extends Component
     {
         return {
             id : this.state.id,
+            customerId : this.props.customerId,
             name : this.state.name,
             content : this.state.content,
-            language : this.state.language,
-            country : this.state.country,
-            logoFile : this.state.logoFile,
-            headerFile : this.state.headerFile,
-            isNew : this.state.isNew
+            languageId : this.state.languageId,
+            countryId : this.state.countryId,
+            type : this.state.type
         }
     }
 
-    takeClasses(classes)
+    putItemToState(item)
     {
-        const result = [ ];
+        const state = this.state;
 
-        for(var key in classes)
-            classes[key] && result.push(key);
+        state.id = parseInt(item.id);
+        state.name = item.name;
+        state.content = item.content;
+        state.languageId  = item.languageId;
+        state.countryId = item.countryId;
+        state.type = state.propsType || item.type;
 
-        return result;
+        this.setState(state);
     }
 
-    getClassesFor(fieldName)
+    clearForm()
     {
-        return this.takeClasses({
-            'form-group' : true,
-            'has-error' : this.state.errors[fieldName]
-        });
+        $(document).ready(() => $('[href="#TemplateForm_Tab1"]').tab('show'));
+
+        this.putItemToState(TemplateForm.emptyFormItem);
+        this.setState({ filesDirectory : null, templateId : '', files : [ ], errors : { } });
+        this.formChanged = false;
     }
 
-    callOnCreate()
+    isNew()
     {
+        return this.state.id > 0;
+    }
+
+    getFormGroupClass(fieldName)
+    {
+        return this.state.errors[fieldName] ? 'form-group has-error' : 'form-group';
+    }
+
+    saveForm()
+    {
+        const i18n = this.context.i18n;
         const item = this.extractItemFromState();
-        const errors = this.validateItem(item);
-    }
+        const errors = this.validateForm(item) || { };
+        const hasErrors = errors && Object.keys(errors).length > 0;
 
-    callOnUpdate()
-    {
-        this.props.onUpdate();
-    }
+        this.setState({ errors : errors });
 
-    callOnCancel()
-    {
-        this.props.onUpdate();
+        if(hasErrors)
+        {
+            this.context.showNotification(i18n.getMessage('TemplateForm.saveForm.notification.error'), 'error');
+            return Promise.resolve(false);
+        }
+        else
+        {
+            const currentId = this.state.id;
+            const currentTemplateId = this.state.templateId;
+            const showSuccess = () => this.context.showNotification(i18n.getMessage('TemplateForm.saveForm.notification.success'), 'success');
+            const showError = (e) => this.context.showNotification(e.message, 'error');
+            const processResponse = (res) =>
+            {
+                const filesDirectory = `${this.state.templateFileDirectory}${res.id}`;
+                let promise;
+
+                if(currentTemplateId)
+                {
+                    const templateFilesDirectory = `${this.state.templateFileDirectory}${currentTemplateId}`;
+                    promise = Api.copyFilesOfTemplate(this.props.customerId, templateFilesDirectory, filesDirectory);
+                }
+                else
+                {
+                    promise = Promise.resolve();
+                }
+
+                return promise.then(() =>
+                {
+                    this.putItemToState(res);
+                    this.setState({ filesDirectory : filesDirectory, templateId : '' });
+                    this.formChanged = false;
+
+                    return res;
+                });
+            }
+
+            let promise = currentId ? Api.updateTemplate(this.props.customerId, currentId, item)
+                : Api.addTemplate(this.props.customerId, item);
+
+            return promise.then(processResponse).then(showSuccess)
+                .delay(500).then(() => this.preview.reload())
+                .then(() => true).catch(showError);
+        }
     }
 
     render()
     {
+        const errorKeys = this.state.errors && Object.keys(this.state.errors);
+        const state = this.state;
+        const { i18n, locale } = this.context;
+        const blobBasePath = `${document.location.origin}/blob/public/api/c_${this.props.customerId}/files`;
+
+        const overwriteFileModalButtons = { 'yes' : i18n.getMessage('System.yes'), 'no' : i18n.getMessage('System.no') };
+
         return(
             <div>
                 <div className="col-md-8 form-horizontal">
-                    <div className={this.getClassesFor('name')}>
-                        <label htmlFor="name" className="col-sm-2 control-label text-left">Template name</label>
-                        <div className="col-sm-1 text-right"></div>
-                        <div className="col-sm-9">
-                            <input type="text" className="form-control col-sm-8" id="name" value={this.state.name} readOnly={!this.state.isNew} onChange={e => this.handleOnChange(e, 'name')} placeholder="Template name" />
-                        </div>
-                    </div>
-                    <div className={this.getClassesFor('content')}>
-                        <label htmlFor="content" className="col-sm-2 control-label text-left">Template content</label>
-                        <div className="col-sm-1 text-right"></div>
-                        <div className="col-sm-9">
-                            <textarea className="form-control" rows="10" id="content" value={this.state.content} onChange={e => this.handleOnChange(e, 'content')}></textarea>
-                        </div>
-                    </div>
-                    <div className="form-group">
-                        <label htmlFor="templateContent" className="col-sm-2 control-label text-left">Template upload</label>
-                        <div className="col-sm-1 text-right"></div>
-                        <div className="col-sm-9">
-                            <button type="button" className="btn btn-default" onClick={() => this.dropzone.open()}>Upload file</button>
-                        </div>
-                    </div>
-                    <div className={this.getClassesFor('language')}>
-                        <label htmlFor="language" className="col-sm-2 control-label text-left">Template language</label>
-                        <div className="col-sm-1 text-right"></div>
-                        <div className="col-sm-9">
-                            <this.LanguageField key='languages' id="language" actionUrl={document.location.origin} value={this.state.language} onChange={e => this.handleOnChange(e, 'language')} />
-                        </div>
-                    </div>
-                    <div className={this.getClassesFor('country')}>
-                        <label htmlFor="country" className="col-sm-2 control-label text-left">Template country</label>
-                        <div className="col-sm-1 text-right"></div>
-                        <div className="col-sm-9">
-                            <this.CountryField key='countries' id="country" actionUrl={document.location.origin} value={this.state.country} onChange={e => this.handleOnChange(e, 'country')} />
-                        </div>
-                    </div>
-                    <div className="form-group">
-                        <label className="col-sm-2 control-label text-left">Logo file</label>
-                        <div className="col-sm-1 text-right"></div>
-                        <div className="col-sm-7">
-                            <input type="text" className="form-control col-sm-8" readOnly={true} value={this.state.logoFile} placeholder="Logo file" />
-                            {
-                                this.state.logoFile && <img src={`/blob/public/api/${this.props.tenantId}/files${this.state.logoFile}`} style={{ maxWidth : 150, maxHeight : 113, marginTop : 10 }} />
-                            }
-                        </div>
-                        <div className="col-sm-2">
-                            <button className="btn btn-default pull-left" onClick={() => this.showFileSelector('logoFile')}><span className="glyphicon glyphicon-folder-open"></span></button>
-                        </div>
-                    </div>
-                    <div className="form-group">
-                        <label className="col-sm-2 control-label text-left">Header file</label>
-                        <div className="col-sm-1 text-right"></div>
-                        <div className="col-sm-7">
-                            <input type="text" className="form-control col-sm-8" readOnly={true} value={this.state.headerFile} placeholder="Header file" />
-                            {
-                                this.state.headerFile && <img src={`/blob/public/api/${this.props.tenantId}/files${this.state.headerFile}`} style={{ maxWidth : 150, maxHeight : 113, marginTop : 10 }} />
-                            }
-                        </div>
-                        <div className="col-sm-2">
-                            <button className="btn btn-default pull-left" onClick={() => this.showFileSelector('headerFile')}><span className="glyphicon glyphicon-folder-open"></span></button>
-                        </div>
-                    </div>
-                    <div className="form-submit text-right">
-                        {
-                            this.props.allowCancel && <button type="submit" className="btn btn-default" onClick={() => this.callOnCancel()}>Cancel</button>
-                        }
-                        {
-                            this.props.allowCreate && this.state.isNew
-                                && <button type="submit" className="btn btn-primary" onClick={() => this.callOnCreate()}>Create</button>
-                        }
-                        {
-                            this.props.allowUpdate && !this.state.isNew
-                                && <button type="submit" className="btn btn-primary" onClick={() => this.callOnUpdate()}>Update</button>
-                        }
-                    </div>
+                  {
+                      errorKeys && errorKeys.length > 0 &&
+                      <div className="alert alert-danger">
+                          <ul>
+                              {errorKeys.map(fieldName => state.errors[fieldName].map(message => <li>{message}</li>))}
+                          </ul>
+                      </div>
+                  }
+                  <div className={this.getFormGroupClass('type')}>
+                      <label htmlFor="type" className="col-sm-2 control-label text-left">{i18n.getMessage('TemplateForm.label.type')}</label>
+                      <div className="col-sm-1 text-right"></div>
+                      <div className="col-sm-9">
+                          <select className="form-control col-sm-8" id="type" disabled={state.type && state.type.length} onChange={e => this.handleOnChange(e, 'type')} defaultValue={state.type}>
+                              <option value=""></option>
+                              <option value="email">{i18n.getMessage('TemplateForm.type.email')}</option>
+                              <option value="landingpage">{i18n.getMessage('TemplateForm.type.landingpage')}</option>
+                          </select>
+                      </div>
+                  </div>
+                  {
+                      this.props.allowCopy &&
+                      <div className={this.getFormGroupClass('template')}>
+                          <label htmlFor="template" className="col-sm-2 control-label text-left">{i18n.getMessage('TemplateForm.label.template')}</label>
+                          <div className="col-sm-1 text-right"></div>
+                          <div className="col-sm-9">
+                              <select className="form-control col-sm-8" id="template" disabled={this.props.templateId} onChange={e => this.applyTemplate(e.target.value)} defaultValue={state.templateId}>
+                                  <option value=""></option>
+                                  {
+                                      state.templates && state.templates.map(template =>
+                                      {
+                                          if(template.id != state.id && template.type === state.type)
+                                              return(<option key={template.id} value={template.id}>{template.name} ({template.languageId || '-'})</option>);
+                                      })
+                                  }
+                              </select>
+                          </div>
+                      </div>
+                  }
+                  <div className={this.getFormGroupClass('name')}>
+                      <label htmlFor="name" className="col-sm-2 control-label text-left">{i18n.getMessage('TemplateForm.label.name')}</label>
+                      <div className="col-sm-1 text-right"></div>
+                      <div className="col-sm-9">
+                          <input type="text" className="form-control col-sm-8" id="name" value={state.name} readOnly={state.id} onChange={e => this.handleOnChange(e, 'name')} />
+                      </div>
+                  </div>
+                  <div className={this.getFormGroupClass('content')}>
+                      <label htmlFor="content" className="col-sm-2 control-label text-left">{i18n.getMessage('TemplateForm.label.content')}</label>
+                      <div className="col-sm-1 text-right"></div>
+                      <div className="col-sm-9">
+                          <textarea className="form-control" rows="10" id="content" value={state.content} onChange={e => this.handleOnChange(e, 'content')}></textarea>
+                      </div>
+                  </div>
+                  <div className="form-group">
+                      <label htmlFor="templateContent" className="col-sm-2 control-label text-left">{i18n.getMessage('TemplateForm.label.uploadFile')}</label>
+                      <div className="col-sm-1 text-right"></div>
+                      <div className="col-sm-9">
+                          <button type="button" className="btn btn-default" onClick={() => this.dropzone.open()}>{i18n.getMessage('TemplateForm.button.upload')}</button>
+                      </div>
+                  </div>
+                  <div className={this.getFormGroupClass('language')}>
+                      <label htmlFor="language" className="col-sm-2 control-label text-left">{i18n.getMessage('TemplateForm.label.language')}</label>
+                      <div className="col-sm-1 text-right"></div>
+                      <div className="col-sm-9">
+                          <this.LanguageField key='languages' id="language" optional={true} value={state.languageId} onChange={e => this.handleOnChange(e, 'languageId')} />
+                      </div>
+                  </div>
+                  <div className={this.getFormGroupClass('country')}>
+                      <label htmlFor="country" className="col-sm-2 control-label text-left">{i18n.getMessage('TemplateForm.label.country')}</label>
+                      <div className="col-sm-1 text-right"></div>
+                      <div className="col-sm-9">
+                          <this.CountryField key='countries' id="country" optional={true} value={state.countryId} onChange={e => this.handleOnChange(e, 'countryId')} />
+                      </div>
+                  </div>
+                  <div className={this.getFormGroupClass('preview')}>
+                      <label htmlFor="country" className="col-sm-2 control-label text-left">{i18n.getMessage('TemplateForm.label.preview')}</label>
+                      <div className="col-sm-1 text-right"></div>
+                      <div className="col-sm-9">
+                          {
+                              (state.id > 0 &&
+                                  <TemplatePreview
+                                      ref={node => this.preview = node}
+                                      templateId={state.id}
+                                      customerId={this.props.customerId}
+                                      allowCopy={false}
+                                      allowEdit={false} />)
+                              ||
+                                  <p className="lead">{i18n.getMessage('TemplateForm.text.previewNotAvailable')}</p>
+                          }
+                      </div>
+                  </div>
                 </div>
                 <div className="col-md-4">
-                    <table className="table">
-                          <thead>
-                                <tr>
-                                      <th>Placeholder</th>
-                                      <th>Description</th>
-                                </tr>
-                          </thead>
-                          <tbody>
-                                {
-                                    this.templateFields.map(field =>
-                                    {
-                                        return(
-                                            <tr key={field.key}>
-                                                  <td>{'{{'+ field.key +'}}'}</td>
-                                                  <td>{field.description}</td>
-                                            </tr>
-                                        )
-                                    })
-                                }
-                          </tbody>
-                    </table>
+                    <h4>{i18n.getMessage('TemplateForm.header.placeholder')}</h4>
                 </div>
+                <div className="col-md-4" style={{ maxHeight : '450px', overflow : 'scroll', wordWrap : 'break-word' }}>
+                    <ul className="list-group">
+                        {
+                            templateFields[locale].map(field =>
+                            {
+                                return(<li key={field.key} className="list-group-item"><strong>{'{{'+ field.key +'}}'}</strong><br /><span>{field.description}</span></li>)
+                            })
+                        }
+                    </ul>
+                </div>
+                <div className="col-md-4">
+                    <h4>{i18n.getMessage('TemplateForm.header.files')}</h4>
+                </div>
+                <div className="col-md-4" style={{ maxHeight : '300px', overflow : 'scroll', wordWrap : 'break-word' }}>
+                    <ul className="list-group">
+                        {
+                            state.files.length && state.files.map((file, index) =>
+                            {
+                                return(
+                                    <li key={file.path} className="list-group-item">
+                                        <div className="row">
+                                            <div className="col-xs-3">
+                                                <img src={`${blobBasePath}${file.path}?inline=true`} className="img-thumbnail" />
+                                            </div>
+                                            <div className="col-xs-9">
+                                                <span className="input-group">
+                                                    <input type="text" className="form-control" id={`file_${index}`} readOnly={true} defaultValue={`${blobBasePath}${file.path}`} />
+                                                    <span className="input-group-btn">
+                                                        <ClipboardButton className="btn btn-default" data-clipboard-target={`#file_${index}`}>
+                                                            <i className="fa fa-clipboard" aria-hidden="true"></i>
+                                                        </ClipboardButton>
+                                                    </span>
+                                                </span>
+                                            </div>
+                                        </div>
+                                    </li>
+                                )
+                            })
+                            || <div><hr /><p className="lead">{i18n.getMessage('TemplateForm.text.noFilesAvailable')}</p></div>
+                        }
+                    </ul>
+                </div>
+
                 <Dropzone style={{ display: 'none' }} ref={node => this.dropzone = node} onDrop={files => this.uploadTemplate(files.shift())} />
+
                 <ModalDialog
-                    title="Choose a file"
-                    visible={this.state.showFileSelector}
-                    buttons={[ 'close' ]}
-                    onButtonClick={() => this.hideFileSelector()}>
-                    <FileManager
-                        tenantId={this.props.tenantId}
-                        selectorVersion={true}
-                        filesDirectory={this.props.filesDirectory}
-                        onFileSelection={item => this.handleFileSelection(this.state.showFileSelector, item.path)}>
-                    </FileManager>
-                </ModalDialog>
-                <ModalDialog
-                    title="Overwrite template file"
-                    message="Do you really want to overwrite the existing template content?"
-                    visible={this.state.showOverwriteTemplateDialog}
-                    buttons={[ 'yes', 'no' ]}
+                    title={i18n.getMessage('TemplateForm.modal.overwriteFile.title')}
+                    message={i18n.getMessage('TemplateForm.modal.overwriteFile.message')}
+                    visible={state.showOverwriteTemplateDialog}
+                    buttons={overwriteFileModalButtons}
                     onButtonClick={button => this.onOverwriteDialogButtonClick(button)}>
                 </ModalDialog>
             </div>
