@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, Legend } from 'recharts';
 import Messages from './i18n';
 import request from 'superagent';
@@ -6,7 +7,7 @@ import request from 'superagent';
 class FunnelChart extends React.Component
 {
     static contextTypes = {
-        i18n: React.PropTypes.object.isRequired,
+        i18n: PropTypes.object.isRequired,
     };
 
     constructor(props)
